pyproject.toml
[project]
name = "fastapi-passkey-auth"
version = "0.1.0"
description = "FastAPI application with passkey authentication support"
authors = [
    {name = "Dennis Lee",email = "wylee2000@gmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi (>=0.115.6,<0.116.0)",
    "webauthn (>=2.4.0,<3.0.0)",
    "pydantic-settings (>=2.7.1,<3.0.0)",
    "python-jose[cryptography] (>=3.3.0,<4.0.0)",
    "aiosqlite (>=0.20.0,<0.21.0)",
    "alembic (>=1.14.0,<2.0.0)",
    "sqlalchemy[asyncio] (>=2.0.37,<3.0.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "pydantic[email] (>=2.10.5,<3.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)"
]

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
